refactor(shared): extract base64url-JSON codec to shared helper
The compact-shim and the messages web-search-shim both round-trip private
payloads through an opaque-string slot on the wire (`encrypted_content`,
`encrypted_index`) using the same base64url(JSON) codec — no envelope or
prefix marker, with foreign upstream blobs detected structurally.

Extract `encodeBase64UrlJson` / `decodeBase64UrlJson` into
`shared/base64url-json.ts` and import from both shims. Structural
validators stay per-shim since each shim's payload shape differs.
